+++ src/Core/EasyOC.Core/Indexes/FreeSqlDocumentIndex.cs
@@ -8,7 +8,7 @@ namespace EasyOC.Core.Indexes
     public abstract class FreeSqlDocumentIndex : MapIndex, IFreeSqlMapDocumentIndex
     {
         [Column(IsPrimary = true, IsIdentity = true, IsNullable = false)]
-        new public virtual int Id { get => base.Id; set { base.Id = value; } }
+        new public virtual long Id { get => base.Id; set { base.Id = value; } }
 
         public virtual int DocumentId { get; set; }
